requirements.txt
azure-functions
numpy